% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capstone.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{clean entire dataset}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{a data frame that contains the imported base}
}
\value{
cleaned dataframe is returned with country names in title case and coordinates as numeric insread of character
}
\description{
This function "cleans" the initial dataset so it can be further used in geoms: creates date column in POSTIXct format with given Year, Month and Day,
as well as convert LONGITUDE and LATITUDE to numeric and clean location name using previous function [func(eq_location_clean)]. Because the cleaning task is
quite specific to the current database, there is no option to choose columns, related to dates and coordinates. Instead, all of them are nested in the body of function
thus, the only arument is the name of \code{data.frame} in which base is stored
}
\examples{
\dontrun{eq_clean_data(database)->cleaned_base}
}
