% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capstone.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Function to generate leaflet popup text}
\usage{
eq_create_label(data, location, magnitude, death)
}
\arguments{
\item{data}{data used}

\item{location}{location}

\item{magnitude}{intensity}

\item{death}{numbet of deaths occured}
}
\description{
Function to automate creating popup labels in leaflet using html tags
}
\examples{
cleaned_base$popup_text<-eq_create_label(cleaned_base, "LOCATION", "EQ_MAG_ML", "DEATHS")
}
