% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capstone.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{clean location in raw data}
\usage{
eq_location_clean(location)
}
\arguments{
\item{location}{column of dataframe or simply a vector with names of locations to be changed}
}
\value{
the vector of Country names in title case is returned
}
\description{
This function is used in order to change name of Location in database to more convinient format: it keeps only Country name in Titlecase
}
\details{
Source code for the project
}
\examples{
\dontrun{eq_location_clean<-(location=database$COUNTRY)}
}
