% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capstone.R
\name{eq_map}
\alias{eq_map}
\title{create interactive plots based on leaflet}
\usage{
eq_map(data, longitude, latitude, annot_col)
}
\arguments{
\item{data}{data used}

\item{longitude}{x-coord}

\item{latitude}{y-cood}

\item{annot_col}{column to extract popup caption from}
}
\description{
Function to plot dots of hurricanes on ineractive maps with popus containing basic information
}
\examples{
\dontrun{eq_map(cleaned_base, "LONGITUDE", "LATITUDE", "popup_text")}
}
