% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capstone.R
\docType{data}
\name{GeomTimeline_label}
\alias{GeomTimeline_label}
\title{It is modification of previous \code{geom_timeline} that can add captions to hurricanes plotted by date occured. Captions can be applyed to all observation,
or to first n observations, having maximum value of specified metric (for example, first 5 with the highest magnitude level)}
\format{An object of class \code{GeomTimeline_label} (inherits from \code{Geom}, \code{ggproto}) of length 7.}
\usage{
GeomTimeline_label
}
\description{
It is modification of previous \code{geom_timeline} that can add captions to hurricanes plotted by date occured. Captions can be applyed to all observation,
or to first n observations, having maximum value of specified metric (for example, first 5 with the highest magnitude level)
}
\examples{
This function is needed to make geom_timeline_label works. Is not called by user directly
}
\keyword{datasets}
