% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capstone.R
\name{draw_key_timeline}
\alias{draw_key_timeline}
\alias{draw_key_timline}
\title{function to identify legend}
\usage{
draw_key_timline(data, params, size)
}
\description{
This is function that is used in ggproto to make custom legend type in new geom_*.
 @inheritParams ggplot2::draw_key_polygon
}
